+++ src/main/java/dooya/see/adapter/webapi/MemberApi.java
@@ -6,6 +6,9 @@ import dooya.see.adapter.webapi.dto.MemberRegisterResponse;
 import dooya.see.application.member.provided.MemberAuth;
 import dooya.see.application.member.provided.MemberRegister;
 import dooya.see.domain.member.*;
+import dooya.see.domain.member.dto.MemberAuthRequest;
+import dooya.see.domain.member.dto.MemberInfoUpdateRequest;
+import dooya.see.domain.member.dto.MemberRegisterRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
