+++ src/main/java/dooya/see/application/member/MemberAuthService.java
@@ -4,6 +4,7 @@ import dooya.see.application.member.provided.MemberAuth;
 import dooya.see.application.member.provided.MemberFinder;
 import dooya.see.application.member.required.TokenManager;
 import dooya.see.domain.member.*;
+import dooya.see.domain.member.dto.MemberAuthRequest;
 import dooya.see.domain.member.exception.AuthenticateException;
 import dooya.see.domain.shared.Email;
 import lombok.RequiredArgsConstructor;
