+++ src/main/java/dooya/see/application/member/provided/MemberAuth.java
@@ -1,7 +1,7 @@
 package dooya.see.application.member.provided;
 
 import dooya.see.domain.member.Member;
-import dooya.see.domain.member.MemberAuthRequest;
+import dooya.see.domain.member.dto.MemberAuthRequest;
 import dooya.see.domain.member.LoginResult;
 
 /**
