+++ src/test/java/dooya/see/adapter/webapi/CommentApiTest.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import dooya.see.adapter.webapi.dto.*;
 import dooya.see.application.member.provided.MemberRegister;
-import dooya.see.domain.member.MemberAuthRequest;
+import dooya.see.domain.member.dto.MemberAuthRequest;
 import dooya.see.domain.member.MemberFixture;
-import dooya.see.domain.member.MemberRegisterRequest;
+import dooya.see.domain.member.dto.MemberRegisterRequest;
 import dooya.see.domain.post.CommentCreateRequest;
 import dooya.see.domain.post.CommentStatus;
 import dooya.see.domain.post.CommentUpdateRequest;
