+++ src/test/java/dooya/see/adapter/webapi/MemberApiTest.java
@@ -8,6 +8,9 @@ import dooya.see.adapter.webapi.dto.MemberRegisterResponse;
 import dooya.see.application.member.provided.MemberRegister;
 import dooya.see.application.member.required.MemberRepository;
 import dooya.see.domain.member.*;
+import dooya.see.domain.member.dto.MemberAuthRequest;
+import dooya.see.domain.member.dto.MemberInfoUpdateRequest;
+import dooya.see.domain.member.dto.MemberRegisterRequest;
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
