+++ src/test/java/dooya/see/adapter/webapi/PostApiTest.java
@@ -6,8 +6,8 @@ import dooya.see.adapter.webapi.dto.MemberAuthResponse;
 import dooya.see.adapter.webapi.dto.PostCreateResponse;
 import dooya.see.adapter.webapi.dto.PostDetailResponse;
 import dooya.see.application.member.provided.MemberRegister;
-import dooya.see.domain.member.MemberAuthRequest;
-import dooya.see.domain.member.MemberRegisterRequest;
+import dooya.see.domain.member.dto.MemberAuthRequest;
+import dooya.see.domain.member.dto.MemberRegisterRequest;
 import dooya.see.domain.post.*;
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.BeforeEach;
