+++ src/test/java/dooya/see/application/member/provided/MemberRegisterTest.java
@@ -2,6 +2,7 @@ package dooya.see.application.member.provided;
 
 import dooya.see.SeeTestConfiguration;
 import dooya.see.domain.member.*;
+import dooya.see.domain.member.dto.MemberInfoUpdateRequest;
 import dooya.see.domain.member.exception.DuplicateEmailException;
 import dooya.see.domain.member.exception.DuplicateProfileException;
 import jakarta.persistence.EntityManager;
